Fix postgresql non-determinite order
Having updated some ducks with age and pond details, we can no longer rely on their retrival order (at least with Postgresql). We should instead target the pond and rank by age so we know we're getting the first two ducks in the pond.

This doesn't affect the idea of the test which is to make sure that ducks added to the end of the list via the maximum value are indeed added to the end of the list.