Added tests
Demonstrates failures around ordering when moving an item to the second to last position in the list when there are consecutive rankings at those positions.

Also demonstrates that when adding a lot of items to the end of the list, we very quickly run out of space between ranks at the end of the list.