Fix the rebalance method
First we mutate the array so that it represents the intended reposition. Then we rebalance all items in the array according to its mutated new ordering. There might be a more efficient way, but this is the easiest to understand from a code readability point of view.

Extended the collision test to add a third item at the same rank which should get rebalanced.