Fix issue with consecutive final two ranks and setting rank to second to last (#130)
* Added test showing rebalancing bug

* Added tests

Demonstrates failures around ordering when moving an item to the second to last position in the list when there are consecutive rankings at those positions.

Also demonstrates that when adding a lot of items to the end of the list, we very quickly run out of space between ranks at the end of the list.

* Fix indentation

* Abstract average rank calculation to a method

`rank_at_average` detects when there is no space between the ranks we want to insert between. If this occurs a rank rebalancing is triggered before attempting to position the item again.

* Add explanation comments for the checks and changes that `rearrange_ranks` does

* Move away from %Q and prefer hashes vs string interpolation for update_all where possible

* A nil rank is already handled in `update_index_from_position` under `when NilClass`

* `finder` already strips out the current item id from the query

* A complete rewrite of `rebalance_ranks`

Rewritten to be more readable.

Guarantees equal gaps between each item in the list and gives priority to the current item if there are collissions. Also handles the case where there might be multiple collissions with the current item's proposed position.

Gaps at the top and bottom of the list are also maintained (away from `MAX_RANK_VALUE` and `MIN_RANK_VALUE`)

Adds the concept of resetting the cached values such as @finder after a rebalance.

* Refactor `finder`

There's no need to use direct AREL in many situations here. I've converted these cases to normal rails scoping methods.

Also simplified some of the logic.

* The `current_order` needs to include all records

* Add pry and AR debug

* Fix the rebalance method

First we mutate the array so that it represents the intended reposition. Then we rebalance all items in the array according to its mutated new ordering. There might be a more efficient way, but this is the easiest to understand from a code readability point of view.

Extended the collision test to add a third item at the same rank which should get rebalanced.

* The new rebancing routine exposed this test to be incorrect.

* Remove support for Rails 3.2

This is due to the usage of `.where.not()`