Publish ActiveSupport notifications on rebalancing (#203)
* Publish ActiveSupport notifications on rebalancing

Record updates to rebalance ranks do not trigger ActiveRecord callbacks.
This makes it difficult to react to these updates (for example, to index
them in a secondary data store). This commit instruments the updates
with ActiveSupport notifications. Subscribed consumers receive an event
for each rearrangement or rebalancing, the payload of which includes the
triggering instance and the scope and with_same options for the ranking,
which can be used to retrieve the affected records.

* Update README formatting

* Drop Ruby 2.4 from test matrix